+++ example/web/scopes_main.dart
@@ -65,6 +65,7 @@ class MyTestClass {
 
   MyTestClass({this.message = 'world'}) {
     myselfField = this;
+    tornOff = toString;
   }
 
   String hello() => message;
@@ -93,13 +94,25 @@ class MyTestClass {
     libraryFunction('abc');
   }
 
-  // ignore: unused_field
   final _privateField = 'a private field';
+
+  // ignore: unused_element
+  String privateMethod(String s) => '$s : $_privateField';
 
   @override
   String toString() => 'A test class with message $message';
 
+  bool equals(Object other) {
+    if (other is MyTestClass) return message == other.hello();
+    return false;
+  }
+
   Function closure = someFunction;
+
+  String Function() tornOff;
 }
 
 Function someFunction() => null;
+
+// ignore: unused_element
+int _libraryPrivateFunction(int a, int b) => a + b;