+++ dwds/CHANGELOG.md
@@ -1,6 +1,9 @@
 ## 0.6.1
 
-- Add `isolateRef` to `Isolate`s `pauseEvent`s.
+ - Add `isolateRef` to `Isolate`s `pauseEvent`s.
+ - Implements `invoke`.
+ - Adds support for VM object IDs for things that don't have Chrome object Ids
+   (e.g. int, double, bool, null).
 
 ## 0.6.0