+++ dwds/CHANGELOG.md
@@ -1,6 +1,7 @@
 ## 0.6.1
 
  - Add `isolateRef` to `Isolate`s `pauseEvent`s.
+ - Depend on the latest `package:vm_service`.
  - Implements `invoke`.
  - Adds support for VM object IDs for things that don't have Chrome object Ids
    (e.g. int, double, bool, null).